% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Ibov_Stocks.R
\name{GetIbovStocks}
\alias{GetIbovStocks}
\title{Function to download the current components of the Ibovespa index from Wikipedia}
\usage{
GetIbovStocks()
}
\value{
A dataframe that includes a column with the list of tickers of companies that belong to the Ibovespa index
}
\description{
This function scrapes the stocks that constitute the SP500 index from the wikipedia page at https://en.wikipedia.org/wiki/List_of_S%26P_500_companies.
}
\examples{
\dontrun{
df.ibov <- GetIbovStocks()
print(df.ibov$tickers)
}
}
